+++ packages/models/src/customer/Customer/Customer.ts
@@ -3,15 +3,35 @@ import { classes } from "polytype";
 import { DataModel } from "../../base/DataModel.js";
 import assertObjectFound from "../../base/assertObjectFound.js";
 import { ReferenceModel } from "../../base/ReferenceModel.js";
-import type { AsyncResourceVariant } from "../../lib/provideReact.js";
-import { provideReact } from "../../lib/provideReact.js";
 import {
-  CustomerListItemData,
+  AsyncResourceVariant,
+  provideReact,
+} from "../../react/provideReact.js";
+import {
   CustomerData,
-  CustomerListQuery,
+  CustomerListItemData,
+  CustomerListQueryData,
+  CustomerUpdateRequestData,
 } from "./types.js";
+import { ListQueryModel } from "../../base/ListQueryModel.js";
+import { ListDataModel } from "../../base/ListDataModel.js";
+import { ServerListQuery } from "../../server/index.js";
+import { ProjectListQuery } from "../../project/index.js";
 
 export class Customer extends ReferenceModel {
+  public readonly servers: ServerListQuery;
+  public readonly projects: ProjectListQuery;
+
+  public constructor(id: string) {
+    super(id);
+    this.servers = new ServerListQuery({
+      customer: this,
+    });
+    this.projects = new ProjectListQuery({
+      customer: this,
+    });
+  }
+
   public static ofId(id: string): Customer {
     return new Customer(id);
   }
@@ -25,13 +45,16 @@ export class Customer extends ReferenceModel {
     },
   );
 
+  public static query(query: CustomerListQueryData = {}) {
+    return new CustomerListQuery(query);
+  }
+
+  /** @deprecated Use query() */
   public static list = provideReact(
     async (
-      query: CustomerListQuery = {},
-    ): Promise<Readonly<Array<CustomerListItem>>> => {
-      const data = await config.behaviors.customer.list(query);
-      return Object.freeze(data.map((d) => new CustomerListItem(d)));
-    },
+      query: CustomerListQueryData = {},
+    ): Promise<Readonly<Array<CustomerListItem>>> =>
+      new CustomerListQuery(query).execute().then((res) => res.items),
   );
 
   public static get = provideReact(
@@ -45,7 +68,16 @@ export class Customer extends ReferenceModel {
   public getDetailed = provideReact(
     () => Customer.get(this.id),
     [this.id],
-  ) as AsyncResourceVariant<CustomerDetailed, []>;
+  ) as AsyncResourceVariant<() => Promise<CustomerDetailed>>;
+
+  public findDetailed = provideReact(
+    () => Customer.find(this.id),
+    [this.id],
+  ) as AsyncResourceVariant<() => Promise<CustomerDetailed | undefined>>;
+
+  public async update(data: CustomerUpdateRequestData): Promise<void> {
+    await config.behaviors.customer.update(this.id, data);
+  }
 }
 
 // Common class for future extension
@@ -73,5 +105,56 @@ export class CustomerListItem extends classes(
 ) {
   public constructor(data: CustomerListItemData) {
     super([data], [data]);
+  }
+}
+
+export class CustomerListQuery extends ListQueryModel<CustomerListQueryData> {
+  public constructor(query: CustomerListQueryData = {}) {
+    super(query);
+  }
+
+  public refine(query: CustomerListQueryData) {
+    return new CustomerListQuery({
+      ...this.query,
+      ...query,
+    });
+  }
+
+  public execute = provideReact(async () => {
+    const { items, totalCount } = await config.behaviors.customer.list({
+      limit: config.defaultPaginationLimit,
+      ...this.query,
+    });
+
+    return new CustomerList(
+      this.query,
+      items.map((d) => new CustomerListItem(d)),
+      totalCount,
+    );
+  }, [this.queryId]);
+
+  public getTotalCount = provideReact(async () => {
+    const { totalCount } = await this.refine({ limit: 1 }).execute();
+    return totalCount;
+  }, [this.queryId]);
+
+  public findOneAndOnly = provideReact(async () => {
+    const { items, totalCount } = await this.refine({ limit: 2 }).execute();
+    if (totalCount === 1) {
+      return items[0];
+    }
+  }, [this.queryId]);
+}
+
+export class CustomerList extends classes(
+  CustomerListQuery,
+  ListDataModel<CustomerListItem>,
+) {
+  public constructor(
+    query: CustomerListQueryData,
+    customers: CustomerListItem[],
+    totalCount: number,
+  ) {
+    super([query], [customers, totalCount]);
   }
 }