+++ packages/models/src/customer/Customer/types.ts
@@ -1,6 +1,6 @@
 import { MittwaldAPIV2 } from "@mittwald/api-client";
 
-export type CustomerListQuery =
+export type CustomerListQueryData =
   MittwaldAPIV2.Paths.V2Customers.Get.Parameters.Query;
 
 export type CustomerData =
@@ -8,3 +8,6 @@ export type CustomerData =
 
 export type CustomerListItemData =
   MittwaldAPIV2.Operations.CustomerListCustomers.ResponseData[number];
+
+export type CustomerUpdateRequestData =
+  MittwaldAPIV2.Paths.V2CustomersCustomerId.Put.Parameters.RequestBody;