+++ packages/models/src/domain/Ingress/Ingress.ts
@@ -3,10 +3,19 @@ import { classes } from "polytype";
 import { DataModel } from "../../base/DataModel.js";
 import assertObjectFound from "../../base/assertObjectFound.js";
 import { ReferenceModel } from "../../base/ReferenceModel.js";
-import type { AsyncResourceVariant } from "../../lib/provideReact.js";
-import { provideReact } from "../../lib/provideReact.js";
-import { IngressListItemData, IngressData, IngressListQuery } from "./types.js";
+import {
+  AsyncResourceVariant,
+  provideReact,
+} from "../../react/provideReact.js";
+import {
+  IngressData,
+  IngressListItemData,
+  IngressListQueryData,
+  IngressListQueryModelData,
+} from "./types.js";
 import { IngressPath } from "../IngressPath/IngressPath.js";
+import { ListQueryModel } from "../../base/ListQueryModel.js";
+import { ListDataModel } from "../../base/ListDataModel.js";
 
 export class Ingress extends ReferenceModel {
   public static ofId(id: string): Ingress {
@@ -17,11 +26,12 @@ export class Ingress extends ReferenceModel {
     return Ingress.ofId(hostname);
   }
 
+  /** @deprecated: use query() or project.ingresses */
   public static list = provideReact(
-    async (query: IngressListQuery = {}): Promise<Array<IngressListItem>> => {
-      const data = await config.behaviors.ingress.list(query);
-      return data.map((d) => new IngressListItem(d));
-    },
+    async (
+      query: IngressListQueryData = {},
+    ): Promise<Readonly<Array<IngressListItem>>> =>
+      new IngressListQuery(query).execute().then((r) => r.items),
   );
 
   public static find = provideReact(
@@ -44,7 +54,12 @@ export class Ingress extends ReferenceModel {
   public getDetailed = provideReact(
     () => Ingress.get(this.id),
     [this.id],
-  ) as AsyncResourceVariant<IngressDetailed, []>;
+  ) as AsyncResourceVariant<() => Promise<IngressDetailed>>;
+
+  public findDetailed = provideReact(
+    () => Ingress.find(this.id),
+    [this.id],
+  ) as AsyncResourceVariant<() => Promise<IngressDetailed | undefined>>;
 }
 
 export class IngressCommon extends classes(
@@ -83,5 +98,73 @@ export class IngressListItem extends classes(
 ) {
   public constructor(data: IngressListItemData) {
     super([data], [data]);
+  }
+}
+
+export class IngressListQuery extends ListQueryModel<IngressListQueryModelData> {
+  public constructor(query: IngressListQueryModelData = {}) {
+    super(query);
+  }
+
+  public refine(query: IngressListQueryModelData = {}) {
+    return new IngressListQuery({
+      ...this.query,
+      ...query,
+    });
+  }
+
+  public execute = provideReact(async () => {
+    const { project, ...query } = this.query;
+    const { items, totalCount } = await config.behaviors.ingress.list({
+      /** @todo: use this code when pagination is supported by API */
+      // limit: config.defaultPaginationLimit,
+      ...query,
+      projectId: project?.id,
+    });
+
+    return new IngressList(
+      this.query,
+      items.map((d) => new IngressListItem(d)),
+      totalCount,
+    );
+  }, [this.queryId]);
+
+  public getTotalCount = provideReact(async () => {
+    /** @todo: use this code when pagination is supported by API */
+    // const { totalCount } = await this.refine({ limit: 1 }).execute();
+    // return totalCount;
+    const { items } = await this.refine({}).execute();
+    return items.length;
+  }, [this.queryId]);
+
+  public findOneAndOnly = provideReact(async () => {
+    /** @todo: use this code when pagination is supported by API */
+    // const { items, totalCount } = await this.refine({ limit: 2 }).execute();
+    // if (totalCount === 1) {
+    //   return items[0];
+    // }
+    const { items, totalCount } = await this.refine({}).execute();
+    if (totalCount === 1) {
+      return items[0];
+    }
+  }, [this.queryId]);
+}
+
+export class IngressList extends classes(
+  IngressListQuery,
+  ListDataModel<IngressListItem>,
+) {
+  public constructor(
+    query: IngressListQueryModelData,
+    ingresses: IngressListItem[],
+    totalCount: number,
+  ) {
+    super([query], [ingresses, totalCount]);
+  }
+
+  public getDefault() {
+    const defaultIngress = this.items.find((i) => i.data.isDefault);
+    assertObjectFound(defaultIngress, IngressListItem, "IngressList");
+    return defaultIngress;
   }
 }