+++ CHANGELOG.md
@@ -3,6 +3,141 @@
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.61.1](https://github.com/mittwald/api-client-js/compare/4.61.0...4.61.1) (2024-09-24)
+
+**Note:** Version bump only for package root
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [4.61.0](https://github.com/mittwald/api-client-js/compare/4.60.1...4.61.0) (2024-09-24)
+
+### Features
+
+- **Models:** export Model invalidation request option
+  ([8418480](https://github.com/mittwald/api-client-js/commit/8418480baed246963353de45aa076e5a6e300188))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [4.60.1](https://github.com/mittwald/api-client-js/compare/4.60.0...4.60.1) (2024-09-23)
+
+### Bug Fixes
+
+- **Models:** fix type in provideReact
+  ([7f29da5](https://github.com/mittwald/api-client-js/commit/7f29da5894a9605870fdf17ea1122c7230be2b65))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [4.60.0](https://github.com/mittwald/api-client-js/compare/4.59.0...4.60.0) (2024-09-23)
+
+### Features
+
+- **Models:** refresh Models by tags (URLs per default)
+  ([fe69d79](https://github.com/mittwald/api-client-js/commit/fe69d79ca3eee0a1eb9c33abaf070ba64d4a6493))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [4.59.0](https://github.com/mittwald/api-client-js/compare/4.58.2...4.59.0) (2024-09-20)
+
+### Features
+
+- **Commons:** introduce options for request functions
+  ([0af347a](https://github.com/mittwald/api-client-js/commit/0af347a9928136395c4e6db2c2238753b50d4fdc))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [4.58.2](https://github.com/mittwald/api-client-js/compare/4.58.1...4.58.2) (2024-09-20)
+
+**Note:** Version bump only for package root
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [4.58.1](https://github.com/mittwald/api-client-js/compare/4.58.0...4.58.1) (2024-09-10)
+
+**Note:** Version bump only for package root
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [4.58.0](https://github.com/mittwald/api-client-js/compare/4.57.0...4.58.0) (2024-09-04)
+
+### Features
+
+- **Client:** re-export \* from commons
+  ([9bd6a8b](https://github.com/mittwald/api-client-js/commit/9bd6a8bd07154763043d2ac235c4f6445cc4f7b3))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [4.57.0](https://github.com/mittwald/api-client-js/compare/4.56.0...4.57.0) (2024-08-30)
+
+### Features
+
+- **Customer:** add update method
+  ([cccdbe6](https://github.com/mittwald/api-client-js/commit/cccdbe6befbaf403cd7008bf98cae2fb26579f23))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [4.56.0](https://github.com/mittwald/api-client-js/compare/4.55.0...4.56.0) (2024-08-22)
+
+### Features
+
+- **Models:** export Query base models
+  ([5c986db](https://github.com/mittwald/api-client-js/commit/5c986db17693373ea341d3c2f33d6d7323619ace))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [4.55.0](https://github.com/mittwald/api-client-js/compare/4.54.0...4.55.0) (2024-08-22)
+
+### Features
+
+- **Models:** introduce Query Models and List Models
+  ([effb4ba](https://github.com/mittwald/api-client-js/commit/effb4ba91fc1a22393ab32093242650ad1d12a1d))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [4.54.0](https://github.com/mittwald/api-client-js/compare/4.53.0...4.54.0) (2024-08-22)
+
+### Features
+
+- **Models:** add findDetailed methods
+  ([2cb1843](https://github.com/mittwald/api-client-js/commit/2cb1843f7f843b9a66cfc500d20292cd0efcdf52))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
 # [4.53.0](https://github.com/mittwald/api-client-js/compare/4.52.0...4.53.0) (2024-08-22)
 
 ### Features